+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.21.1...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.21.2...HEAD)
+
+## 0.21.2
+
+[Full diff](https://github.com/sider/runners/compare/0.21.1...0.21.2)
 
 - More kindful error message on invalid configuration [#818](https://github.com/sider/runners/pull/818)
 - Add raw content of config file to trace [#819](https://github.com/sider/runners/pull/819)